refactor: mover todos los models al modulo compartido
