+++ src/boe/boe.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
+import { Xml2jsonService } from '../compartido/xml2json.service';
 import { BoeController } from './boe.controller';
 import { BoeService } from './boe.service';
-import { Xml2jsonService } from './xml2json.service';
 
 @Module({
   controllers: [BoeController],
