+++ src/compartido/compartido.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { BoeApiService } from './boe-api.service';
+
+@Module({
+  providers: [BoeApiService],
+})
+export class CompartidoModule {}
