+++ src/compartido/mappers/anuncio.mapper.ts
@@ -1,6 +1,6 @@
-import { Documento } from '@boe/api-models';
-import { construirDetallesDeContrato, constuirContrato, Contrato } from '@contratos/models';
 import { fechaPublicacionMapper, tituloMapper } from '@utils/utils';
+import { Documento } from 'src/compartido/api-models';
+import { construirDetallesDeContrato, constuirContrato, Contrato } from 'src/compartido/models';
 import { beneficiarioMapper } from './beneficiario.mapper';
 import { descripcionMapper } from './descripcion.mapper';
 import { institucionMapper } from './institucion.mapper';
