+++ src/compartido/models/contrato.model.ts
