+++ src/compartido/models/index.ts
