+++ src/utils/utils.ts
@@ -1,5 +1,5 @@
-import { constuirLote, Lote } from '@contratos/models';
-import { Descripcion, ListaDeDefinicion, Termino, Texto } from '../boe/api-models';
+import { constuirLote, Lote } from 'src/compartido/models';
+import { Descripcion, ListaDeDefinicion, Termino, Texto } from '../compartido/api-models';
 
 type ExtractorIndice = (lotes: Termino | Termino[], numLote: Termino) => number;
