+++ test/mocks/ts/beneficiario-anidado.ts
@@ -1,4 +1,4 @@
-import { ListaDeDefinicion } from '@boe/api-models';
+import { ListaDeDefinicion } from 'src/compartido/api-models';
 
 export const mockDeBeneficiarioAnidado: ListaDeDefinicion = {
   dd: [
