+++ test/mocks/ts/contrato-plano-mapeado.ts
@@ -1,4 +1,4 @@
-import { Contrato } from '@contratos/models';
+import { Contrato } from 'src/compartido/models';
 
 export const contratoPlanoMapeado: Contrato = {
   detalles: {
